[cordova] use spacedodge instead of gallery
- It's fine to pack spacedodge sample app

Impacted tests(approved): new 0, update 10, delete 1
Unit test platform: Crosswalk Project for Android 16.44.406.0
Unit test result summary: pass 10, fail 0, block 0

BUG=https://crosswalk-project.org/jira/browse/XWALK-5122
