+++ cordova/cordova-sampleapp-android-tests/sampleapp/Crosswalk_Cordova_SampleApp_Spacedodge_AppFunction.html
@@ -1,6 +1,6 @@
 <!DOCTYPE html>
 <!--
-Copyright (c) 2013 Intel Corporation.
+Copyright (c) 2015 Intel Corporation.
 
 Redistribution and use in source and binary forms, with or without modification,
 are permitted provided that the following conditions are met:
@@ -34,33 +34,21 @@ Authors:
 <html>
   <head>
     <meta charset='utf-8'>
-    <title>Crosswalk_Cordova_SampleApp_Gallery_AppFunction</title>
+    <title>Crosswalk_Cordova_SampleApp_Spacedodge_AppFunction</title>
     <link rel="author" title="Intel" href="http://www.intel.com">
   </head>
   <body>
     <p>
-      <strong>Preconditions:</strong>
-    </p>
-    <ol>
-      <li>Need a good network that can access google server</li>
-    </ol>
-    <p>
       <strong>Test steps:</strong>
     </p>
     <ol>
-      <li>Check the display of "Carousel image gallery" from Gallery UI</li>
-      <li>Check the display of "Carousel video gallery" from Gallery UI</li>
-      <li>Check the display of "Lightbox image gallery" from Gallery UI</li>
-      <li>Check every button from Gallery UI</li>
+      <li>Check every button from spacedodge UI.</li>
     </ol>
     <p>
       <strong>Expected Output:</strong>
     </p>
     <ol>
-      <li>"Carousel image gallery" should display image with auto slides show in section 1 , click next/previous button to navigate to different image</li>
-      <li>"Carousel video gallery" should display video preview in slides show, could play video, could click next/previous button to navigation</li>
-      <li>"Lightbox image gallery" should display image thumbnail in lightbox, could display orignial image when click 1 thumbnail</li>
-      <li>All button and function works well</li>
+      <li>All button and function work well</li>
     </ol>
   </body>
 </html>
