+++ cordova/cordova-sampleapp-android-tests/sampleapp/Crosswalk_Cordova_SampleApp_Spacedodge_Icon.html
@@ -1,6 +1,6 @@
 <!DOCTYPE html>
 <!--
-Copyright (c) 2013 Intel Corporation.
+Copyright (c) 2015 Intel Corporation.
 
 Redistribution and use in source and binary forms, with or without modification,
 are permitted provided that the following conditions are met:
@@ -33,7 +33,7 @@ Authors:
 <html>
   <head>
     <meta charset='utf-8'>
-    <title>Crosswalk_Cordova_SampleApp_Gallery_Orientation</title>
+    <title>Crosswalk_Cordova_SampleApp_Spacedodge_Icon</title>
     <link rel="author" title="Intel" href="http://www.intel.com">
   </head>
   <body>
@@ -41,16 +41,12 @@ Authors:
       <strong>Test steps:</strong>
     </p>
     <ol>
-      <li>Press home button to background Gallery</li>
-      <li>Rotate device from portrait orientation to landscape orientation</li>
-      <li>Rotate device from landscape orientation to portrait orientation</li>
+      <li>Check the Spacedodge's icon in app launcher</li>
     </ol>
     <p>
       <strong>Expected Output:</strong>
     </p>
     <ol>
-      <li>Task manager is able to open</li>
-      <li>The device display with landscape orientation</li>
-      <li>The device display with portrait orientation</li>
+      <li>The icon display correctly and the same as icon in Apps</li>
   </body>
 </html>
