+++ cordova/cordova-sampleapp-android-tests/sampleapp/Crosswalk_Cordova_SampleApp_Spacedodge_Orientation.html
@@ -1,6 +1,6 @@
 <!DOCTYPE html>
 <!--
-Copyright (c) 2013 Intel Corporation.
+Copyright (c) 2015 Intel Corporation.
 
 Redistribution and use in source and binary forms, with or without modification,
 are permitted provided that the following conditions are met:
@@ -33,7 +33,7 @@ Authors:
 <html>
   <head>
     <meta charset='utf-8'>
-    <title>Crosswalk_Cordova_SampleApp_Gallery_Switch</title>
+    <title>Crosswalk_Cordova_SampleApp_Spacedodge_Orientation</title>
     <link rel="author" title="Intel" href="http://www.intel.com">
   </head>
   <body>
@@ -41,16 +41,16 @@ Authors:
       <strong>Test steps:</strong>
     </p>
     <ol>
-      <li>Press home button to background Gallery</li>
-      <li>Long press home button to open task manager</li>
-      <li>Tap Gallery icon to switch back to Gallery</li>
+      <li>Press home button to background Spacedodge</li>
+      <li>Rotate device from portrait orientation to landscape orientation</li>
+      <li>Rotate device from landscape orientation to portrait orientation</li>
     </ol>
     <p>
       <strong>Expected Output:</strong>
     </p>
     <ol>
-      <li>Gallery is background</li>
-      <li>Task manager is able to open </li>
-      <li>Gallery is able to switch back</li>
+      <li>Task manager is able to open</li>
+      <li>The device display with landscape orientation</li>
+      <li>The device display with portrait orientation</li>
   </body>
 </html>
