+++ cordova/cordova-sampleapp-android-tests/sampleapp/Crosswalk_Cordova_SampleApp_Spacedodge_Switch.html
@@ -1,6 +1,6 @@
 <!DOCTYPE html>
 <!--
-Copyright (c) 2013 Intel Corporation.
+Copyright (c) 2015 Intel Corporation.
 
 Redistribution and use in source and binary forms, with or without modification,
 are permitted provided that the following conditions are met:
@@ -33,7 +33,7 @@ Authors:
 <html>
   <head>
     <meta charset='utf-8'>
-    <title>Crosswalk_Cordova_SampleApp_Gallery_Icon</title>
+    <title>Crosswalk_Cordova_SampleApp_Spacedodge_Switch</title>
     <link rel="author" title="Intel" href="http://www.intel.com">
   </head>
   <body>
@@ -41,12 +41,16 @@ Authors:
       <strong>Test steps:</strong>
     </p>
     <ol>
-      <li>Check the Gallery's icon in app launcher</li>
+      <li>Press home button to background Spacedodge</li>
+      <li>Long press home button to open task manager</li>
+      <li>Tap Spacedodge icon to switch back to Spacedodge</li>
     </ol>
     <p>
       <strong>Expected Output:</strong>
     </p>
     <ol>
-      <li>The icon display correctly and the same as icon in Apps</li>
+      <li>Spacedodge is background</li>
+      <li>Task manager is able to open </li>
+      <li>Spacedodge is able to switch back</li>
   </body>
 </html>
