+++ cordova/cordova-sampleapp-android-tests/sampleapp/spacedodge_build.py
@@ -31,26 +31,19 @@
 import unittest
 import os
 import commands
+import glob
 import comm
-import time
 
 
-class TestGalleryAppUninstallWithAppRunning(unittest.TestCase):
+class TestSpacedodgeAppBuild(unittest.TestCase):
 
-    def test_uninstall_withAppRunning(self):
+    def test_build(self):
         comm.setUp()
-        app_name = "gallery"
-        pkg_name = "com.example." + app_name.lower()
-
-        if not comm.check_app_installed(pkg_name, self):
-            comm.app_install(app_name, pkg_name, self)
-
-        # Make sure the app is running
-        comm.app_launch(app_name, pkg_name, self)
-        time.sleep(2)
-
-        # Uninstall the app
-        comm.app_uninstall(pkg_name, self)
+        app_name = "spacedodge"
+        pkg_name = " com.example." + app_name.lower()
+        sample_src_pref = "/tmp/crosswalk-demos/space-dodge-game/base"
+        comm.create(app_name, pkg_name, comm.MODE, sample_src_pref, None, self)
+        comm.build(app_name, False, self)
 
 if __name__ == '__main__':
     unittest.main()
